Modals
No longer require use of .hide. We updated the utility classes to use !important and to avoid clashing, removed the .hide class from modals. Now display: none; is set directly in the modals.less file. When upgrading, be sure to remove the .hide class from existing modals.
Redesign the modals to be mobile-first.
As BS3 is mobile-first, this new modal also starts at the mobile level and scales up via media queries.
Reintroduces .modal-open on the body (so we can nuke the scroll there)
Adds a couple extra levels of markup (namely .modal-dialog and .modal-content) so we can scroll the entire modal rather than overflow a section within the modal.
Related, .modal is now the wrapper, and .modal-content is the modal itself. This is so we can still use position: fixed;, but make the modal relatively positioned so that scrolling moves the entire modal, not something with it.
Added a .modal-title for more consistent and useful targeting of the heading content (previously this was just an <h3> and selector performance wise that sucked).
Test on iOS5. Not sure if we need to go older, but I want to check here to see how well fixed is supported.
Test on Android devices (including native browser and Chrome). I only have a year-old Nexus lying around, so might need help tracking down bugs on older devices.
Add option to programmatically set width of the modal.
Prevent <body> scrollbar and shifting content with overflow: hidden;.

See https://github.com/twbs/bootstrap/pull/6342