[WIP] Bootstrap4 (#744)
* Add bootstrap v4.1.0 and fontawesome

* Change bootstrap-less into bootstrap-sass

Bootstrap v4 uses sass preprocessor instead of less, therefore
in order to move to v4, we have to abandon less and use libsass instead.

Most of the changes are just renaming existing classes/files and so.
The most significant change is moving sources url resolution from
LessUrlSource to UrlImporter. Importer is feature of libsass and allows
to customize nested imports. We use this to resolve webjars, classpath,
servlet context paths and so on.

UrlImporter also supports different ways of importing sass files:
* with and without .sass and .scss extensions
* partials starting with underscore
* css files

Importers don't track nested imports, so we have to track
them manually in order to cache them properly.
We do this in TrackingImporter which creates new SassSources
for imported urls and stores them in collection, which can be used after
successful compilation.

* Add popper.js dependency resource reference.