Bootstrap4 themes (#749)
I updated bootstrap-theme, bootswatch and material design themes to bootstrap 4.
The biggest change is that oficial bootstrap 4 does not provide `bootstrap-theme` as separate stylesheet rather we have to enable two properties in sass source and compile it. I added shell script that automates this process, similarly to scripts in other themes.

Remaining themes are in different state of migration:

* M8tro - No sign of upgrade to 4 (https://github.com/idleberg/m8tro-bootstrap/issues/4)
* Vegibit - No sign of upgrade to 4
* Google - Upgrade in progress (https://github.com/todc/todc-bootstrap/issues/172)

What do you think is the best approach here? Can we just delete them? Or should we wait until todc releases new version of Google theme before we merge this to upstream branch?