Feature/707 font awesome 5 (#790)
I'm somewhat unsure what would be the best way to implement font-awesome-5.x.

This might eventually fix #707.

First thing, changing the version from 4.7 to 5.7 is a breaking change, as (at least some) of the icons have different names or are not available anymore.

Further, font-awesome now has 4 different styles of icons, 3 of which can be used in wicket-bootstrap.

* style `solid`: All icons are free and can be used in the project. To be used with `fas` (or `fa`)
* style `regular`: Only some of those icons are free, many are not. To be used with `far`
* style `brand`: all free. To be used with `fab`
* style `light` all of which are not free - and thus don't need to be further discussed here.

The icons of style `brand` have distinctive names and are not an issue. The icons that are freely available in style `regular` use the same name as those in `solid`.

I have separated the three different free icon styles within FontAwesomeIconTypeBuilder. One enum each `FontAwesomeSolid`, `FontAwesomeRegular` and `FontAwesomeBrand` (all implementing `FontAwesomeGraphic` which is now an interface) define each icon. So as with the additional attributes of the icons (spin, fixedWith etc.), the user has full freedom on how to use which style in which variant.

For the pre-built instances of `FontAwesomeIconType` I see two general approaches (with sub-options) how we can handle those in wicket-bootstrap:

* Aproach 1: Every style is implemented in it's own instance of `FontAwesomeIconType`. We can use tags to distinguish the style (between regular and solid, no name clash for brands)
  * Option 1.1: Use the tag `_r`/`_s` where we have both a variant of style solid and regular (e.g. `address_book_s` and `address_book_r`). Where we only have a solid variant right now, we don't apply the tag `s`. the icon `adjust` currently is only (freely) available as solid. If in the future it will also have a regular variant that is free, we would add `adjust_r` but would also need to rename `adjust` to `adjust_s` for consistency.
    * **Pro:** Some of the icons with only a solid variant will get the same name with font-awesome-5.7 as with font-awesome-4.7. Less migration required for the users of wicket-bootstrap with the version bump.
    * **Cons:** A new icon variant will break the name of the pre-existin solid variant.
  * Option 1.2: Use tags `_s` (solid) or `_r` (regular) with *all* icons of those style, regardless of whether we currently have a solid and a regular (free) variant or not. `address_book_s` and `address_book_r`, but also `adjust_s` (with no `adjust_r`).
   * **Pro:** Once an icon of style regular will become free in a later version of font-awesome, we will only have to add that regular variant `adjust_r` without having to bother about the previously existing solid variant, which is already called `adjust_s`.
   * **Cons**: *All* icons will have a different name with the version change compared to the icon names in wicket-boostrap with font-awesome 4.7.0.
 
I strongly prefer option 2 and have already implemented this pattern within this pull request. It will require more migration work initially for users of wicket-bootstrap, but only once. Later amendments will be simpler.

I was also thinking about a completely separate approach with enhancing the prebuilt instances of `FontAwesomeIconType` to handle both styles (regular, solid) if both are available. If not, we could throw exception if somebody wants to use a non-existing style of an icon. I have the feeling that this would be more aligned with font-awesomes approach with the separate styles. But I'm not sure if and how this could be implemented to play along nicely with the usage of IconType implementations. It would certainly be much more complex to implement. Not sure if it's worthwhile.

References:

* https://fontawesome.com/how-to-use/on-the-web/setup/upgrading-from-version-4
* https://fontawesome.com/how-to-use/on-the-web/referencing-icons/basic-use

Note: While the documentation states that regular style requires a pro license, the detail page about icons lists some icons of the regular style as pro (e.g. https://fontawesome.com/icons/ad?style=solid), some as not (https://fontawesome.com/icons/address-book?style=solid). I have only implemented the regular icons that are listed as free (https://fontawesome.com/icons?d=gallery&m=free).

What do you think?