fix active state of tabs
In Bootstrap 4, active tabs set the .active class on the <a>-Tag
instead of the <li>-Tag (This seems to be Wicket's default behavior).
This commit removes the .active class from the <li>-Tag and adds it
to the <a>-tag in order to fix the styling of active tabs.

For more information, see https://getbootstrap.com/docs/4.0/components/navs/#tabs